requirements/base.txt
Django==3.2.13
asgiref==3.5.2
attrs==21.4.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.4
iniconfig==1.1.1
model-bakery==1.5.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest-django==4.5.2
pytest==7.1.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
zipp==3.8.0
